Convert instance_created_file resource to a data resource
A data resource is the ideal terraform object that represents an
externally created file.
