+++ lxd/provider.go
@@ -198,7 +198,6 @@ func Provider() *schema.Provider {
 			"lxd_container_file":          resourceLxdContainerFile(),
 			"lxd_instance":                resourceLxdInstance(),
 			"lxd_instance_file":           resourceLxdInstanceFile(),
-			"lxd_instance_created_file":   resourceLxdInstanceCreatedFile(),
 			"lxd_network":                 resourceLxdNetwork(),
 			"lxd_profile":                 resourceLxdProfile(),
 			"lxd_project":                 resourceLxdProject(),
@@ -207,6 +206,10 @@ func Provider() *schema.Provider {
 			"lxd_volume":                  resourceLxdVolume(),
 			"lxd_volume_copy":             resourceLxdVolumeCopy(),
 			"lxd_volume_container_attach": resourceLxdVolumeContainerAttach(),
+		},
+
+		DataSourcesMap: map[string]*schema.Resource{
+			"lxd_instance_file": dataLxdInstanceFile(),
 		},
 
 		ConfigureFunc: providerConfigure,
